# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gas Monitor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /refill-orders:
    get:
      operationId: getRefillOrders
      tags: [refill-orders]
      summary: List refill orders (homeowner sees own; supplier sees assigned)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RefillOrder"
    post:
      operationId: createRefillOrder
      tags: [refill-orders]
      summary: Place a new refill order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefillOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefillOrder"
  /refill-orders/{id}:
    get:
      operationId: getRefillOrder
      tags: [refill-orders]
      summary: Get refill order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefillOrder"
    patch:
      operationId: updateRefillOrder
      tags: [refill-orders]
      summary: Update refill order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefillOrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefillOrder"

  # ── Messages ──────────────────────────────────────────────────────────
...
  /supplier/customers:
    get:
      operationId: getSupplierCustomers
      tags: [supplier]
      summary: List linked homeowner customers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierCustomer"
    post:
      operationId: linkCustomer
      tags: [supplier]
      summary: Link a homeowner as a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierCustomer"
  /supplier/customers/{id}:
    delete:
      operationId: unlinkCustomer
      tags: [supplier]
      summary: Remove a customer link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /supplier/inventory:
    get:
      operationId: getInventory
      tags: [supplier]
      summary: Get supplier inventory
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [supplier]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /supplier/inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [supplier]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /supplier/dispatches:
    get:
      operationId: getDispatches
      tags: [supplier]
      summary: List dispatches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dispatch"
    post:
      operationId: createDispatch
      tags: [supplier]
      summary: Create a dispatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /supplier/dispatches/{id}:
    patch:
      operationId: updateDispatch
      tags: [supplier]
      summary: Update dispatch status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
...
  /supplier/analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [supplier]
      summary: Revenue analytics by period
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: monthly
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"

  # ── Homeowner ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RefillOrder:
      type: object
      required: [id, homeownerId, status, deliveryType, volumeGallons, totalPrice, createdAt]
      properties:
        id:               { type: integer }
        homeownerId:      { type: integer }
        supplierId:       { type: ["integer", "null"] }
        status:           { type: string, enum: [pending, dispatched, en_route, delivered, cancelled] }
        deliveryType:     { type: string, enum: [standard, express, emergency] }
        volumeGallons:    { type: number }
        totalPrice:       { type: number }
        deliveryFee:      { type: number }
        notes:            { type: ["string", "null"] }
        estimatedArrival: { type: ["string", "null"] }
        driverName:       { type: ["string", "null"] }
        createdAt:        { type: string, format: date-time }
    RefillOrderInput:
      type: object
      required: [deliveryType, volumeGallons]
      properties:
        supplierId:    { type: integer }
        deliveryType:  { type: string, enum: [standard, express, emergency] }
        volumeGallons: { type: number }
        notes:         { type: string }
    RefillOrderUpdate:
      type: object
      properties:
        status:           { type: string, enum: [pending, dispatched, en_route, delivered, cancelled] }
        supplierId:       { type: integer }
        estimatedArrival: { type: string }
        driverName:       { type: string }

    # ── Messages ──────────────────────────────────────────────────────
...
    SupplierCustomer:
      type: object
      required: [id, supplierId, homeownerId, homeownerName, homeownerEmail, linkedAt]
      properties:
        id:             { type: integer }
        supplierId:     { type: integer }
        homeownerId:    { type: integer }
        homeownerName:  { type: string }
        homeownerEmail: { type: string }
        gasLevelPercent: { type: ["number", "null"] }
        deviceCount:    { type: integer }
        linkedAt:       { type: string, format: date-time }
    CustomerLinkInput:
      type: object
      required: [homeownerEmail]
      properties:
        homeownerEmail: { type: string }
    InventoryItem:
      type: object
      required: [id, supplierId, cylinderType, quantityAvailable, updatedAt]
      properties:
        id:                { type: integer }
        supplierId:        { type: integer }
        cylinderType:      { type: string }
        quantityAvailable: { type: integer }
        pricePerUnit:      { type: number }
        updatedAt:         { type: string, format: date-time }
    InventoryItemInput:
      type: object
      required: [cylinderType, quantityAvailable]
      properties:
        cylinderType:      { type: string }
        quantityAvailable: { type: integer }
        pricePerUnit:      { type: number }
    InventoryItemUpdate:
      type: object
      properties:
        quantityAvailable: { type: integer }
        pricePerUnit:      { type: number }
    Dispatch:
      type: object
      required: [id, supplierId, orderId, status, createdAt]
      properties:
        id:          { type: integer }
        supplierId:  { type: integer }
        orderId:     { type: integer }
        driverName:  { type: ["string", "null"] }
        truckNumber: { type: ["string", "null"] }
        status:      { type: string, enum: [pending, en_route, delivered, cancelled] }
        notes:       { type: ["string", "null"] }
        customerName: { type: ["string", "null"] }
        address:     { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }
    DispatchInput:
      type: object
      required: [orderId]
      properties:
        orderId:     { type: integer }
        driverName:  { type: string }
        truckNumber: { type: string }
        notes:       { type: string }
    DispatchUpdate:
      type: object
      properties:
        status:      { type: string, enum: [pending, en_route, delivered, cancelled] }
        driverName:  { type: string }
        truckNumber: { type: string }
        notes:       { type: string }
...
    RevenueDataPoint:
      type: object
      required: [label, revenue, orders]
      properties:
        label:   { type: string }
        revenue: { type: number }
        orders:  { type: integer }
...