# lib/api-zod/src/generated/types/dispatchInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gas Monitor API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DispatchInput {
  orderId: number;
  driverName?: string;
  truckNumber?: string;
  notes?: string;
}