# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gas Monitor API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  quantityAvailable?: number;
  pricePerUnit?: number;
}